+++ src/pages/kangjimin/Main/JmMain.js
@@ -1,27 +1,20 @@
-import React, { useState } from 'react';
-import CommentList from './Comments';
+import React, { useState, useEffect } from 'react';
 import AsideFooter from './Aside';
+import Feed from './Feed';
 import './Main.scss';
 import '../../../styles/reset.scss';
 import '../../../styles/common.scss';
 
 function JmMain() {
-  const [commentValue, setCommentValue] = useState('');
+  const [feedList, setFeedList] = useState([]);
 
-  const [commentArray, setCommentArray] = useState([]);
-  const submitComment = () => {
-    setCommentArray([...commentArray, commentValue]);
-    //...스프레드연산자 공부
-    setCommentValue('');
-  };
-
-  // const submitComment = e => {
-  //   e.preventDefault();
-  //   commentArray.push(commentValue);
-  //   // setCommentValue('');
-  // };
-
-  const activeSubmit = commentValue.length > 0;
+  useEffect(() => {
+    fetch('/data/feedData.json', { method: 'GET' })
+      .then(res => res.json())
+      .then(data => {
+        setFeedList(data);
+      });
+  });
 
   return (
     <>
@@ -60,83 +53,9 @@ function JmMain() {
       <div className="main">
         <div className="feeds">
           <article>
-            <div className="user">
-              <img
-                className="profile"
-                src="/images/kangjimin/cat2.jpg"
-                alt=""
-              />
-              <span className="id">mingming</span>
-              <img
-                src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/more.png"
-                alt=""
-              />
-            </div>
-            <div className="feedImg">
-              <img src="/images/kangjimin/cat1.jpg" alt="" />
-            </div>
-            <div className="icons">
-              <img
-                src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/heart.png"
-                alt=""
-              />
-              <img
-                src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/comment.png"
-                alt=""
-              />
-              <img
-                src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/share.png"
-                alt=""
-              />
-              <img
-                src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/bookmark.png"
-                alt=""
-              />
-            </div>
-
-            <div className="text">
-              <p className="like">
-                <img
-                  className="profile"
-                  src="/images/kangjimin/cat3.jpg"
-                  alt=""
-                />
-                <span className="bold">jmmmmmmmm</span>님&nbsp;
-                <span className="bold">여러 명</span>이 좋아합니다
-              </p>
-              <p>
-                <span className="id">mingming</span>고양이 사진보며 힐링...
-                <span className="view">더 보기</span>
-              </p>
-              <ul className="comments">
-                <li>
-                  <span className="id">neceosecius</span>
-                  <p>거봐 좋았잖아~~~~</p>
-                  <img
-                    src="https://s3.ap-northeast-2.amazonaws.com/cdn.wecode.co.kr/bearu/heart.png"
-                    alt=""
-                  />
-                </li>
-                <CommentList commentArray={commentArray} />
-              </ul>
-              <p>42분 전</p>
-            </div>
-
-            <form className="comment-box" onSumbit={submitComment}>
-              <input
-                type="text"
-                placeholder="댓글 달기..."
-                value={commentValue}
-                onChange={event => setCommentValue(event.target.value)}
-              />
-              <button
-                className={activeSubmit ? 'activeSub' : 'nonActiveSub'}
-                disabled={!activeSubmit}
-                onClick={submitComment}
-              >
-                게시
-              </button>
-            </form>
+            {feedList.map(feed => (
+              <Feed key={feed.id} feed={feed} />
+            ))}
           </article>
         </div>
